+++ iris-mpc-cpu/src/lib.rs
@@ -3,8 +3,10 @@ pub mod genesis;
 pub mod hawkers;
 pub mod hnsw;
 pub mod network;
-#[rustfmt::skip]
-pub(crate) mod proto_generated;
 pub mod protocol;
 pub mod py_bindings;
 pub mod shares;
+pub mod utils;
+
+#[rustfmt::skip]
+pub(crate) mod proto_generated;
