+++ iris-mpc-cpu/src/utils/constants.rs
@@ -1,0 +1,1 @@
+pub const N_PARTIES: usize = 3;
